13978 - Remove unnecessary dependency on most recent stomp versions
The new stomp gem defines various helpful Stomp::Error::* classes which
help us handle various types of failure better where in the past it
was all just the same kind of exception.

In adding those to rescue blocks we inadvertantly forced that everyone
use the newest gem which is not what we want.

This commit detects older gems and create dummy constants for these new
errors which will never be used with older gems and effectively noops
those rescue blocks.

This is temporary code at best, soon as the 1.2.x series Stomp gem is
wider tested we will require at least 1.2.2.