14239 - RPC::Client#compound_filter does not correctly add to options
The filters in the main protocol hash are all arrays and calls to
Client#compound_filter and Client#fact_filter etc should append to those
arrays.  Compound filter did not and so setting the filter from the RPC
client just did not work.

We now append to the array instead