14496 - ActiveMQ :connect_timeout should default to 30 to allow IPv4 fallback to work in multi homed nodes
Prior to this the default connect timeout was 5 seconds, this means if
the underlying socket does not get established within 5 seconds the
attempt will be stopped and the next in the pool will be tried.

In the case of a dual homed host where the IPv6 address is not reachable
but the IPv4 one is the underlying socket connection will first try the
v6 address and after some time try the v4 address.

By increasing the initial connect timeout we allow this v6 to v4
fallback to happen before timing out the attempt.

Also added this setting to the default stomp connector