14255 - Discovery method should be pluggable
A proof of concept that makes discovery pluggable it includes 3 plugins
one thats the traditional broadcast, one that discovers against a
flatfile and one that uses mongodb and the mongo registration agent.

Discovery plugins need DDLs, in the DDL files they declare their
capabilities - can they do fact filter etc. and mcollective will
validate the filter prior to calling the plugin.

RPC clients will force to directed mode for anything but the mc plugin

You'd elect to use a different default discovery mode by setting
default_discovery_mode = flatfile in your client config file

The plugin application will show help for these plugins, it can package
them as rpm/deb and all the usual stuff

The RPC library will adjust discovery time based on the timeout in the
DDL for the plugin unless a custom timeout is specified on the CLI

In a case where compound filters are used we force the 'mc' discovery
plugin since data queries means this is the only thing that makes sense

All application now has a --nodes option that can take a flat file with
a list of nodes

There are a few issues here:

 - tests are failing and new tests are needed