14874 - Setting the collective in an rpc client should reset the client
When changing collectives the list of nodes applicable to a specific
filter will change.

Prior to this commit setting the collective did not clear the discovered
host list which might leave a stale expectation in the client leading to
unexpected failures.

We now reset the client after changing the collective, we also validate
the collective exist before changing it.