15191 - Use UUIDs for request IDs rather than our hacky way
Previously we created our own request ids, these were not great multiple
requests in the same milli second on the same host could cause request
id clashes...not a problem in functionality with how the new message
flow works but it would be a auditing problem to have dupes.

We now use a more reliably random UUID implementation based on OpenSSL
128 bit random numbers and RFC compliant UUID formats

The SSL#uuid function returns standard format UUIDs for either a given
string or based on OpenSSL::BN.rand().  In the Message class we remove
the -s in the UUID thus keeping new and old request ids the same size
and format ensuring everyone who may have audit parsers wont break as a
result of this code