15273 - add an option to force display => failed behavior to rpc app
Since introducing summaries in the DDL it seems it would be common to
only ever want to see the summaries for some sets of results.

A package status query for example might scroll for pages when really
all you want is the summary.

This patch adds a --display flag that takes ok, all or failed as options
which maps to :ok, :always and :failed in the DDL display field