15348 - aggregate functions should have access to the full reply not just the data section
While first designing the aggregate interface we thought aggregate
functions would only need access to the data and not information like
reply status or sender.

In exploring what kind of functions can be written we decided to write a
function that detects outliers in result sets and found we'd want to
report the outliers by name and data.

This commit makes the full reply available to the aggregate functions so
they can figure out hostnames and status etc as well