15557 - input validation should be done on the server too
We recently started requiring the DDL on the agents but did not do much
past using their metadata.

On the clients requests are validated against the DDL prior to sending
it across the wire but no matching validation is done on the agents
unless users code specific validate logic into their agents.

Prior to doing authorization we now validate the incoming request
against the DDL, this will validate all the inputs against the
:validation clauses in the DDL files.

This means the validation is consistant on both the client and the
server and will make agents easier to write.  We will make the various
validation modes pluggable and expose those into the DDL soon thus
improving the quality and types of validation thats possible in the DDL.