15602 - discovery timeout being incorrectly calculated
Prior to this commit the logic to calculate discovery timeout were
spread all over the code.  This is a legacy throwback from before any
one piece of code was responsible for discovery.

We recently added a discovery plugin subsystem that codifies discovery
behavior but the discovery timeout calculations were not completely
moved into this subsystem.

This commit moves all the discovery claculations into the M::Discovery
class and everywhere else that needs to show or calculate discovery time
is using this single piece of logic to calculate the timeouts needed for
any given request