15793 - display of summaries should provide better feedback when something went wrong
Prior to this the absense of aggregate results would simply result in a
block of empty space in the output of printrpcstats.  This was confusing
to the user.

We now assume that any aggregate result #to_s that returns "" were not
capable of calculating results for some reason and show a friendly error
message