18466 - Data::Result#method_missing raises incorrect exception
When calling a method on a class that does not exist ruby expects a
NoMethodError to be raised.

The Result::Data class has a method_missing that allows access to data
items via faked methods and it checks for unknown data items and raises
an error but unfortunately it raised the wrong exception type.

We now raise NoMethodError as expected