19366 - error code collision on PLMC18 in ddl/base.rb
There was an inadvertant error code collision that resulted in a failure
when trying to interopolate the data.

Ideally we'd somehow test for this - no idea how to go about it though.

We could in theory on interprolation failure always retry once using
the default supplied arguments but this would lead to further confusion
as we would not know what code to show the user then, it's better to
just fail