15182 - Support Stomp 1.1 in the RabbitMQ and ActiveMQ plugins
A common user problem is that idle Stomp connections get expired
by session tracking firewalls and NAT devices.

Stomp protocol version 1.1 supports combating this with protocol
level heartbeats.

The Stomp gem added support for this in 1.2.9 but it was broken,
the upcoming 1.2.10 gem will support these properly in combination
with recent ActiveMQ and RabbitMQ version.

This commit adds:

  plugin.activemq.heartbeat_interval
  plugin.activemq.stomp_1_0_fallback
  plugin.activemq.max_hbrlck_fails
  plugin.activemq.max_hbread_fails

and the rabbit equivelants as well as extra logging.